refactor: reorganize Neo4j tool structure and add dotenv support
